requirements.txt
# Core dependencies
httpx>=0.25.0
python-telegram-bot>=20.7
rapidfuzz>=3.5.0
pydantic>=2.5.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

